feat: make list a generic component, fix: reorganize files
