+++ src/components/List/List.js
@@ -1,0 +1,27 @@
+import React from "react";
+import ListItem from "./ListItem";
+
+class List extends React.Component {
+  render() {
+    const { items, isShoppingCart, handleClickItem } = this.props;
+    console.log(items);
+    const listItems = items
+      ? items.map(item => (
+          <ListItem
+            key={item.id}
+            item={item}
+            readonly={isShoppingCart}
+            clickItem={handleClickItem}
+          />
+        ))
+      : " ";
+
+    return (
+      <React.Fragment>
+        <ul>{listItems}</ul>
+      </React.Fragment>
+    );
+  }
+}
+
+export default List;
